moved entities to each directory
