+++ src/main/java/com/programmers/coffeeorder/controller/OrderController.java
@@ -1,9 +1,9 @@
 package com.programmers.coffeeorder.controller;
 
 import com.programmers.coffeeorder.controller.bind.CoffeeOrderSubmit;
-import com.programmers.coffeeorder.entity.CoffeeOrder;
-import com.programmers.coffeeorder.entity.CoffeeProduct;
-import com.programmers.coffeeorder.entity.CoffeeProductOrderItem;
+import com.programmers.coffeeorder.entity.order.CoffeeOrder;
+import com.programmers.coffeeorder.entity.product.coffee.CoffeeProduct;
+import com.programmers.coffeeorder.entity.product.coffee.CoffeeProductOrderItem;
 import com.programmers.coffeeorder.service.order.CoffeeOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
