+++ src/main/java/com/programmers/coffeeorder/entity/order/OrderStatus.java
@@ -1,15 +1,15 @@
-package com.programmers.coffeeorder.entity;
+package com.programmers.coffeeorder.entity.order;
 
 public enum OrderStatus {
-    NOT_DELIVERED,
-    DELIVERING,
-    DELIVERED;
+    CREATED,
+    CANCELLED,
+    ACCEPTED;
 
     public static OrderStatus of(String input) {
         try {
             return OrderStatus.valueOf(input.toUpperCase());
         } catch (IllegalArgumentException ex) {
-            return OrderStatus.NOT_DELIVERED;
+            return OrderStatus.CREATED;
         }
     }
 }
