+++ src/main/java/com/programmers/coffeeorder/service/order/CoffeeOrderService.java
@@ -1,6 +1,6 @@
 package com.programmers.coffeeorder.service.order;
 
-import com.programmers.coffeeorder.entity.CoffeeOrder;
+import com.programmers.coffeeorder.entity.order.CoffeeOrder;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
