+++ src/main/java/com/programmers/coffeeorder/service/product/CoffeeProductService.java
@@ -1,6 +1,6 @@
 package com.programmers.coffeeorder.service.product;
 
-import com.programmers.coffeeorder.entity.CoffeeProduct;
+import com.programmers.coffeeorder.entity.product.coffee.CoffeeProduct;
 
 import java.util.List;
